feat: implement form parsing and routing enhancements with new internal modules
