+++ nexios/http/request.py
@@ -10,9 +10,8 @@ from nexios._utils.async_helpers import (
     AwaitableOrContextManagerWrapper,
 )
 from nexios.session.base import BaseSessionInterface
-from nexios.http.formparsers import FormParser, MultiPartParser
 from nexios.structs import URL, Address, FormData, Headers, QueryParams, State
-from .formparsers import FormParser, MultiPartException, MultiPartParser
+from nexios._internals._formparsers import FormParser, MultiPartException, MultiPartParser
 
 try:
     from python_multipart.multipart import parse_options_header  # type:ignore
